+++ src/calendar.js
@@ -1,7 +1,8 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import ReactDOM from 'react-dom'
 
-import get from 'lodash/get';
+import get from 'lodash/get'
 
 import Moment from 'moment'
 import 'moment-range'
@@ -11,18 +12,16 @@ import { Levels, Units } from './constants'
 import Navigation from './nav'
 import Cell from './cell'
 
-
 class Calendar extends Component {
-
   constructor(props) {
     super(props)
 
     this.state = {
-      windowHeight: window.innerHeight
+      windowHeight: window.innerHeight,
     }
   }
 
-  static PropTypes = {
+  static propTypes = {
     datetime: PropTypes.object.isRequired,
     onSelect: PropTypes.func.isRequired,
     level: PropTypes.string.isRequired,
@@ -92,8 +91,7 @@ class Calendar extends Component {
     const lvl = Levels[this.props.level]
     if (Moment(this.props.datetime).isSame(Moment(), 'day')) {
       this.props.onSelect(Moment(), !lvl.down)
-    }
-    else {
+    } else {
       this.props.onSelect(Moment())
     }
   }
@@ -106,15 +104,13 @@ class Calendar extends Component {
         const end = datetime.clone().add(7, 'years')
         let years = []
 
-        Moment()
-          .range(start, end)
-          .by(Units.YEAR, year => {
-            years.push({
-              label: year.format('YYYY'),
-              selected: year.isSame(datetime, 'year'),
-            })
+        Moment().range(start, end).by(Units.YEAR, year => {
+          years.push({
+            label: year.format('YYYY'),
+            selected: year.isSame(datetime, 'year'),
           })
-        return [years[0].label, years[years.length-1].label].join('-')
+        })
+        return [years[0].label, years[years.length - 1].label].join('-')
 
       case 'months':
         return datetime.format('YYYY')
@@ -125,15 +121,12 @@ class Calendar extends Component {
       case 'hours':
         return null
     }
-
   }
 
   getCells(unit, datetime) {
     datetime = datetime || Moment()
 
-    const type = unit === 'hours' && this.props.timeStep
-      ? 'minutes'
-      : unit
+    const type = unit === 'hours' && this.props.timeStep ? 'minutes' : unit
 
     switch (type) {
       case 'years': {
@@ -142,15 +135,13 @@ class Calendar extends Component {
         let years = []
         const format = get(this.props, 'options.format.year') || 'YYYY'
 
-        Moment()
-          .range(start, end)
-          .by(Units.YEAR, year => {
-            years.push({
-              moment: year,
-              label: year.format(format),
-              selected: year.isSame(datetime, 'year'),
-            })
+        Moment().range(start, end).by(Units.YEAR, year => {
+          years.push({
+            moment: year,
+            label: year.format(format),
+            selected: year.isSame(datetime, 'year'),
           })
+        })
 
         return years
       }
@@ -161,15 +152,13 @@ class Calendar extends Component {
         let months = []
         const format = get(this.props, 'options.format.month') || 'MMM'
 
-        Moment()
-          .range(start, end)
-          .by(Units.MONTH, month => {
-            months.push({
-              moment: month,
-              label: month.format(format),
-              selected: month.isSame(datetime, 'month'),
-            })
+        Moment().range(start, end).by(Units.MONTH, month => {
+          months.push({
+            moment: month,
+            label: month.format(format),
+            selected: month.isSame(datetime, 'month'),
           })
+        })
 
         return months
       }
@@ -180,26 +169,23 @@ class Calendar extends Component {
         let days = []
         const format = get(this.props, 'options.format.day') || 'D'
 
-        Moment.weekdaysMin()
-          .forEach(day => {
-            days.push({
-              label: day,
-              header: true,
-            })
+        Moment.weekdaysMin().forEach(day => {
+          days.push({
+            label: day,
+            header: true,
           })
+        })
 
-        Moment()
-          .range(start, end)
-          .by(Units.DAY, day => {
-            days.push({
-              moment: day,
-              label: day.format(format),
-              past: day.isBefore(datetime, 'month'),
-              future: day.isAfter(datetime, 'month'),
-              selected: day.isSame(datetime, 'day'),
-              today: day.isSame(Moment(), 'day'),
-            })
+        Moment().range(start, end).by(Units.DAY, day => {
+          days.push({
+            moment: day,
+            label: day.format(format),
+            past: day.isBefore(datetime, 'month'),
+            future: day.isAfter(datetime, 'month'),
+            selected: day.isSame(datetime, 'day'),
+            today: day.isSame(Moment(), 'day'),
           })
+        })
 
         return days
       }
@@ -210,27 +196,25 @@ class Calendar extends Component {
         let hours = []
         const closeBefore = datetime.clone().subtract(31, 'minutes')
         const closeAfter = datetime.clone().add(31, 'minutes')
-        const format = get(this.props, 'options.format.hour') || 'h:mm a'
+        const format = get(this.props, 'options.format.hour') || 'HH:mm'
 
-        Moment()
-          .range(start, end)
-          .by(Units.HOUR, hour => {
-            hours.push({
-              moment: hour,
-              label: hour.format(format),
-              selected: hour.isSame(datetime, 'minute'),
-              nearestBefore: hour.isBetween(closeBefore, datetime),
-              nearestAfter: hour.isBetween(datetime, closeAfter),
-            })
-            let halfHour = hour.clone().add(30, 'minutes')
-            hours.push({
-              moment: halfHour,
-              label: halfHour.format(format),
-              selected: halfHour.isSame(datetime, 'minute'),
-              nearestBefore: halfHour.isBetween(closeBefore, datetime),
-              nearestAfter: halfHour.isBetween(datetime, closeAfter),
-            })
+        Moment().range(start, end).by(Units.HOUR, hour => {
+          hours.push({
+            moment: hour,
+            label: hour.format(format),
+            selected: hour.isSame(datetime, 'minute'),
+            nearestBefore: hour.isBetween(closeBefore, datetime),
+            nearestAfter: hour.isBetween(datetime, closeAfter),
           })
+          let halfHour = hour.clone().add(30, 'minutes')
+          hours.push({
+            moment: halfHour,
+            label: halfHour.format(format),
+            selected: halfHour.isSame(datetime, 'minute'),
+            nearestBefore: halfHour.isBetween(closeBefore, datetime),
+            nearestAfter: halfHour.isBetween(datetime, closeAfter),
+          })
+        })
 
         return hours
       }
@@ -239,37 +223,40 @@ class Calendar extends Component {
         const start = datetime.clone().startOf('day')
         const end = datetime.clone().endOf('day')
         let minutes = []
-        const format = get(this.props, 'options.format.minute') || 'HH:mm'
+        const format = get(this.props, 'options.format.hour') || 'HH:mm'
 
-        Moment()
-          .range(start, end)
-          .by(Units.MINUTE, (minute) => {
-            const _minutes = minute.minutes()
+        Moment().range(start, end).by(Units.MINUTE, minute => {
+          const _minutes = minute.minutes()
 
-            if (_minutes === 0) {
-              minutes.push({
-                moment: minute,
-                label: minute.format(format),
-                selected: minute.isSame(datetime, 'minute')
-              })
-            }
-            else if ((_minutes % this.props.timeStep) === 0) {
-              minutes.push({
-                moment: minute,
-                label: minute.format(format),
-                selected: minute.isSame(datetime, 'minute')
-              })
-            }
-          })
+          if (_minutes === 0) {
+            minutes.push({
+              moment: minute,
+              label: minute.format(format),
+              selected: minute.isSame(datetime, 'minute'),
+            })
+          } else if (_minutes % this.props.timeStep === 0) {
+            minutes.push({
+              moment: minute,
+              label: minute.format(format),
+              selected: minute.isSame(datetime, 'minute'),
+            })
+          }
+        })
 
         return minutes
       }
-
     }
   }
 
   render() {
-    const { level, datetime, theme, inputRect, hideOutsideDateTimes } = this.props
+    const {
+      level,
+      datetime,
+      theme,
+      classes,
+      inputRect,
+      hideOutsideDateTimes,
+    } = this.props
 
     let calendarClass = theme.calendarBelow;
 
@@ -279,13 +266,14 @@ class Calendar extends Component {
 
     return (
       <div
-        className={calendarClass}
+        className={cn(this.props.className, calendarClass)}
         onMouseDown={e => this.props.above(true)}
         onMouseUp={e => this.props.above(false)}
+        style={this.props.style}
       >
-        { level != 'hours' &&
+        {level != 'hours' &&
           <Navigation
-            id={this.props.id}
+            instance={this.props.instance}
             onPrev={::this.onNavigateLeft}
             onNext={::this.onNavigateRight}
             onTitle={::this.onNavigateUp}
@@ -310,14 +298,18 @@ class Calendar extends Component {
                   break
               }
 
-              if (level === 'hours' && hideOutsideDateTimes && !this.props.validate(cell.moment, level)) {
+              if (
+                level === 'hours' &&
+                hideOutsideDateTimes &&
+                !this.props.validate(cell.moment, level)
+              ) {
                 return null
               }
 
               return (
                 <Cell
                   key={i}
-                  ref={(cell.selected || cell.nearestBefore) ? 'selected' : null}
+                  ref={cell.selected || cell.nearestBefore ? 'selected' : null}
                   label={cell.label}
                   level={level}
                   type={type}
@@ -329,18 +321,15 @@ class Calendar extends Component {
                   invalid={this.props.validate(cell.moment, level)}
                 />
               )
-            }).filter( cell => cell != null )
-          }
-          { level != 'hours' &&
-            <div className={theme.today} onClick={::this.onToday}>
-              { get(this.props, 'options.format.today') || 'Today' }
-            </div>
-          }
+            }).filter(cell => cell != null)}
+          {level != 'hours' &&
+            <div className={classes.today} onClick={::this.onToday}>
+              {get(this.props, 'options.format.today') || 'Today'}
+            </div>}
         </div>
       </div>
     )
   }
-
 }